add filters to the map and bug fixes regarding the population
